+++ small_text/stopping_criteria/__init__.py
@@ -1,2 +1,10 @@
 from small_text.stopping_criteria.base import DeltaFScore
+from small_text.stopping_criteria.change import ClassificationChange
 from small_text.stopping_criteria.kappa import KappaAverage
+
+
+__all__ = [
+    'ClassificationChange',
+    'DeltaFScore',
+    'KappaAverage'
+]